Updated curl command for better reliability
The curl command used to download a specific zip file has been updated. The new version includes retry parameters, which will attempt the download 5 times with a delay of 10 seconds between each try. This change improves the reliability of the script in case of network issues or temporary unavailability of the file.